useMetronome is now a thing, but it still needs some work.
